+++ vcxproj2cmake/PathUtils.cs
@@ -1,3 +1,4 @@
+using System.Text;
 using System.Text.RegularExpressions;
 
 namespace vcxproj2cmake;
@@ -41,5 +42,79 @@ static class PathUtils
             normalizedPath = normalizedPath[..^1];
 
         return normalizedPath;
+    }
+
+    public static string[] SplitArguments(string arguments)
+    {
+        var result = new List<string>();
+        var current = new StringBuilder();
+        var inQuotes = false;
+        var argumentStarted = false;
+
+        for (int i = 0; i < arguments.Length; i++)
+        {
+            var c = arguments[i];
+
+            if ((c == ' ' || c == '\t') && !inQuotes)
+            {
+                if (argumentStarted)
+                {
+                    result.Add(current.ToString());
+                    current.Clear();
+                    argumentStarted = false;
+                }
+
+                continue;
+            }
+
+            argumentStarted = true;
+
+            if (c == '\\')
+            {
+                var backslashCount = 0;
+
+                while (i < arguments.Length && arguments[i] == '\\')
+                {
+                    backslashCount++;
+                    i++;
+                }
+
+                if (i == arguments.Length || arguments[i] != '"')
+                {
+                    current.Append('\\', backslashCount);
+                    i--;
+                    continue;
+                }
+
+                current.Append('\\', backslashCount / 2);
+
+                if (backslashCount % 2 == 0)
+                    inQuotes = !inQuotes;
+                else
+                    current.Append('"');
+
+                continue;
+            }
+
+            if (c == '"')
+            {
+                if (inQuotes && i + 1 < arguments.Length && arguments[i + 1] == '"')
+                {
+                    current.Append('"');
+                    i++;
+                }
+                else                
+                    inQuotes = !inQuotes;                
+
+                continue;
+            }
+
+            current.Append(c);
+        }
+
+        if (argumentStarted)
+            result.Add(current.ToString());
+
+        return result.ToArray();
     }
 }